+++ core/models.py
@@ -1,7 +1,16 @@
 import hashlib
-from django.db import models
 
+from modelcluster.models import ClusterableModel, ParentalKey
+from wagtail.admin.edit_handlers import FieldPanel, InlinePanel, MultiFieldPanel, PageChooserPanel, StreamFieldPanel
+from wagtail.core import blocks
+from wagtail.core.fields import StreamField
+from wagtail.core.models import Orderable, Page
 from wagtail.images.models import Image, AbstractImage, AbstractRendition
+from wagtail_personalisation.blocks import PersonalisedStructBlock
+from wagtail_personalisation.models import PersonalisablePageMixin
+from wagtail.snippets.models import register_snippet
+
+from django.db import models
 
 
 class AbstractObjectHash(models.Model):
@@ -50,3 +59,81 @@ class Rendition(AbstractRendition):
 
     class Meta:
         unique_together = (('image', 'filter_spec', 'focal_point_key'))
+
+
+@register_snippet
+class Tour(ClusterableModel):
+    page = models.OneToOneField('wagtailcore.Page', on_delete=models.CASCADE, related_name='tour')
+    title = models.CharField(max_length=255)
+    body = models.CharField(max_length=255)
+    button_text = models.CharField(max_length=255)
+
+    panels = [
+        PageChooserPanel('page'),
+        FieldPanel('title'),
+        FieldPanel('body'),
+        FieldPanel('button_text'),
+        MultiFieldPanel([InlinePanel('steps')], heading='Steps'),
+    ]
+
+    def __str__(self):
+        return self.page.title
+
+
+class TourStep(Orderable):
+    title = models.CharField(max_length=255)
+    body = models.CharField(max_length=255)
+    position = models.CharField(max_length=255)
+    selector = models.CharField(max_length=255)
+    tour = ParentalKey(Tour, on_delete=models.CASCADE, related_name='steps')
+
+    panels = [
+        FieldPanel('title'),
+        FieldPanel('body'),
+        FieldPanel('position'),
+        FieldPanel('selector'),
+    ]
+
+
+@register_snippet
+class Product(models.Model):
+    name = models.CharField(max_length=255)
+
+    panels = [
+        FieldPanel('name'),
+    ]
+
+    def __str__(self):
+        return self.name
+
+
+@register_snippet
+class Country(models.Model):
+    name = models.CharField(max_length=255)
+
+    panels = [
+        FieldPanel('name'),
+    ]
+
+    def __str__(self):
+        return self.name
+
+    class Meta:
+        verbose_name_plural = 'Countries'
+
+
+class PersonalisedPage(PersonalisablePageMixin, Page):
+
+    body = StreamField([
+        (
+            'body', PersonalisedStructBlock(
+                [('paragraph', blocks.RichTextBlock())],
+                template='core/personalised_page_struct_block.html',
+                icon='pilcrow'
+            )
+        )
+    ])
+
+    content_panels = Page.content_panels + [
+        StreamFieldPanel('body'),
+    ]