+++ sso/backends.py
@@ -1,5 +1,7 @@
 import directory_sso_api_client.backends
 
+from django.templatetags.static import static
+
 from sso.models import BusinessSSOUser
 
 
@@ -13,3 +15,12 @@ class BusinessSSOUserBackend(directory_sso_api_client.backends.SSOUserBackend):
         parsed = response.json()
         user_kwargs = self.user_kwargs(session_id=session_id, parsed=parsed)
         return BusinessSSOUser(**user_kwargs)
+
+    def user_kwargs(self, session_id, parsed):
+        kwargs = super().user_kwargs(session_id=session_id, parsed=parsed)
+        user_profile = parsed.get('user_profile') or {}
+        kwargs['profile_image'] = user_profile.get(
+            'profile_image',
+            static('images/user-icon.png')
+        )
+        return kwargs